Use NewPgn as Pgn
